+++ js/const.js
@@ -1,0 +1,48 @@
+const PHOTOS_COUNT = 25;
+
+const MinMaxValues = {
+  LIKES_MIN: 15,
+  LIKES_MAX: 200,
+  AVATAR_MIN: 1,
+  AVATAR_MAX: 6,
+  COMMENTS_COUNT_MIN: 0,
+  COMMENTS_COUNT_MAX: 30,
+  COMMENTS_ID_MIN: 1,
+  COMMENTS_ID_MAX: 1000,
+};
+
+const USERS_NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const DESCRIPTIONS = [
+  'котик',
+  'собачка',
+  'птичка',
+  'закат на море',
+  'озеро в лесу',
+  'караван в пустыне',
+  'у бабушки на даче',
+  'город в дожде',
+  'зимняя прогулка',
+  'поездка на велосипедах',
+  'мои друзья',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {PHOTOS_COUNT, MinMaxValues, USERS_NAMES, DESCRIPTIONS, COMMENTS};
